+++ cmd/beacon-chain/flags/BUILD.bazel
@@ -12,6 +12,7 @@ go_library(
     visibility = [
         "//beacon-chain:__subpackages__",
         "//cmd/beacon-chain:__subpackages__",
+        "//shared/gateway:__pkg__",
     ],
     deps = [
         "//shared/cmd:go_default_library",
