+++ shared/gateway/BUILD.bazel
@@ -1,22 +1,23 @@
 # gazelle:ignore
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
+load("@io_bazel_rules_go//go:def.bzl", "go_test")
+load("@prysm//tools/go:def.bzl", "go_library")
 
 go_library(
     name = "go_default_library",
     srcs = [
-        "cors.go",
         "gateway.go",
-        "handlers.go",
         "log.go",
     ],
-    importpath = "github.com/prysmaticlabs/prysm/beacon-chain/gateway",
+    importpath = "github.com/prysmaticlabs/prysm/shared/gateway",
     visibility = [
-        "//beacon-chain/gateway/server:__pkg__",
-        "//beacon-chain/node:__pkg__",
+        "//beacon-chain:__subpackages__",
+        "//validator:__subpackages__",
     ],
     deps = [
         "//proto/beacon/rpc/v1:go_grpc_gateway_library",
+        "//proto/validator/accounts/v2:ethereum_validator_account_gateway_proto",
         "//shared:go_default_library",
+        "//validator/web:go_default_library",
         "@com_github_grpc_ecosystem_grpc_gateway//runtime:go_default_library",
         "@com_github_pkg_errors//:go_default_library",
         "@com_github_prysmaticlabs_ethereumapis//eth/v1alpha1:go_grpc_gateway_library",
@@ -25,5 +26,17 @@ go_library(
         "@org_golang_google_grpc//:go_default_library",
         "@org_golang_google_grpc//connectivity:go_default_library",
         "@org_golang_google_grpc//credentials:go_default_library",
+    ],
+)
+
+go_test(
+    name = "go_default_test",
+    srcs = ["gateway_test.go"],
+    embed = [":go_default_library"],
+    deps = [
+        "//cmd/beacon-chain/flags:go_default_library",
+        "//shared/testutil/require:go_default_library",
+        "@com_github_sirupsen_logrus//hooks/test:go_default_library",
+        "@com_github_urfave_cli_v2//:go_default_library",
     ],
 )
