+++ validator/node/BUILD.bazel
@@ -37,6 +37,7 @@ go_library(
         "//shared/event:go_default_library",
         "//shared/featureconfig:go_default_library",
         "//shared/fileutil:go_default_library",
+        "//shared/gateway:go_default_library",
         "//shared/params:go_default_library",
         "//shared/prereq:go_default_library",
         "//shared/prometheus:go_default_library",
@@ -50,7 +51,6 @@ go_library(
         "//validator/keymanager:go_default_library",
         "//validator/keymanager/imported:go_default_library",
         "//validator/rpc:go_default_library",
-        "//validator/rpc/gateway:go_default_library",
         "//validator/slashing-protection:go_default_library",
         "//validator/slashing-protection/iface:go_default_library",
         "@com_github_pkg_errors//:go_default_library",
