--- validator/rpc/gateway/BUILD.bazel
@@ -1,20 +1,0 @@
-load("@prysm//tools/go:def.bzl", "go_library")
-
-go_library(
-    name = "go_default_library",
-    srcs = [
-        "gateway.go",
-        "log.go",
-    ],
-    importpath = "github.com/prysmaticlabs/prysm/validator/rpc/gateway",
-    visibility = ["//validator:__subpackages__"],
-    deps = [
-        "//proto/validator/accounts/v2:ethereum_validator_account_gateway_proto",
-        "//validator/web:go_default_library",
-        "@com_github_grpc_ecosystem_grpc_gateway//runtime:go_default_library",
-        "@com_github_pkg_errors//:go_default_library",
-        "@com_github_rs_cors//:go_default_library",
-        "@com_github_sirupsen_logrus//:go_default_library",
-        "@org_golang_google_grpc//:go_default_library",
-    ],
-)
